move cleaner processor
